+++ Source/Lab1/OptimizationContexts/BoundedOptimizationContext.cs
@@ -2,12 +2,12 @@
 
 public class BoundedOptimizationContext : IOptimizationContext
 {
-    public BoundedOptimizationContext(decimal a, decimal b)
+    public BoundedOptimizationContext(double a, double b)
     {
         A = a;
         B = b;
     }
 
-    public decimal A { get; }
-    public decimal B { get; }
+    public double A { get; }
+    public double B { get; }
 }
